# scenarios/ambiguous/authoring/implement_T3/files/src/main/java/dev/rajeev/shortener/domain/CreateLinkRequest.java
package dev.rajeev.shortener.domain;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;

/**
 * Create request. Bean Validation annotations document the contract and fail fast at the edge;
 * {@link LinkService} re-validates explicitly so the rules hold for every caller (not only HTTP).
 */
public record CreateLinkRequest(
        @NotBlank @Size(max = 8192) String url,
        @Pattern(regexp = LinkRules.CUSTOM_ALIAS_REGEX, message = "alias must be 6-32 URL-safe characters") String customAlias,
        @Size(max = 128) String ownerId) {}